+++ slack_app_constructs_cdk/slack_app_oauth_constructs_stack.py
@@ -1,15 +1,15 @@
 from aws_cdk import CfnParameter, Duration, RemovalPolicy, Stack
 from aws_cdk import aws_apigateway as apigw_
+from aws_cdk import aws_dynamodb as ddb_
 from aws_cdk import aws_iam as iam_
 from aws_cdk import aws_lambda as lambda_
 from aws_cdk.aws_logs import LogGroup, RetentionDays
 from constructs import Construct
 
 lambda_dir = "lambda"
-PARAMETER_NAME = "/apps/slack_app/k_cdk/token"
-SLACK_COMMAND = "/testcdk"
+CLIENT_ID_PARAMETER_NAME = "/apps/slack_app/k_cdk/client_id"
+CLIENT_SECRET_PARAMETER_NAME = "/apps/slack_app/k_cdk/client_secret"
 
-get_team_domains = lambda settings: list(settings.keys())
 get_team_ids = lambda settings: [v["team_id"] for v in settings.values() if v.get("team_id")]
 
 
@@ -21,7 +21,7 @@ def get_channel_ids(settings):
     return ret
 
 
-class SlackAppConstructsStack(Stack):
+class SlackAppOAuthConstructsStack(Stack):
     def __init__(self, scope: Construct, id: str, settings, **kwargs) -> None:
         super().__init__(scope, id, **kwargs)
         self.id = id
@@ -34,31 +34,31 @@ class SlackAppConstructsStack(Stack):
             type="String",
         ).value_as_string
 
-        # Create function AsyncWorker
-        self.func_async_worker = self.create_lambda("AsyncWorker", custom_role=None)
+        table_name = f"{id}Table"
 
-        # Create function SyncWorker
-        self.func_sync_worker = self.create_lambda("SyncWorker", custom_role=None)
+        # Create a dynamodb table
+        table = self.create_dynamodb_table(table_name)
 
-        # Create function and role for ImmediateResponse
-        func_immediate_response_role = self.create_immediate_response_execution_role(f"{id}-ImmediateResponse")
-        func_immediate_response = self.create_lambda("ImmediateResponse", custom_role=func_immediate_response_role)
-        func_immediate_response.add_environment("SlackAppTokenParameterKey", PARAMETER_NAME)
-        func_immediate_response.add_environment("SlackCommand", SLACK_COMMAND)
-        func_immediate_response.add_environment("AsyncWorkerLambdaFunctionName", f"{id}-AsyncWorker")
-        func_immediate_response.add_environment("SyncWorkerLambdaFunctionName", f"{id}-SyncWorker")
-        func_immediate_response.add_environment("SlackChannelIds", ",".join(get_channel_ids(settings)))
-        func_immediate_response.add_environment("SlackDomains", ",".join(get_team_domains(settings)))
-        func_immediate_response.add_environment("SlackTeamIds", ",".join(get_team_ids(settings)))
-
+        # Create function and role for OAuth
+        func_oauth_role = self.create_func_oauth_execution_role(f"{id}-OAuth", table_arn=table.table_arn)
+        func_oauth = self.create_lambda("OAuth", custom_role=func_oauth_role)
+        func_oauth.add_environment("SlackAppClientIdParameterKey", CLIENT_ID_PARAMETER_NAME)
+        func_oauth.add_environment("SlackAppClientSecretParameterKey", CLIENT_SECRET_PARAMETER_NAME)
+        func_oauth.add_environment("SlackAppOAuthDynamoDBTable", table_name)
+        func_oauth.add_environment("SlackChannelIds", ",".join(get_channel_ids(settings)))
+        func_oauth.add_environment("SlackTeamIds", ",".join(get_team_ids(settings)))
 
         api = apigw_.LambdaRestApi(
             self, f"{id}-API",
             description=f"{id} API",
-            endpoint_configuration=apigw_.EndpointConfiguration(types=[apigw_.EndpointType.EDGE]),
-            handler=func_immediate_response,
+            endpoint_configuration=apigw_.EndpointConfiguration(types=[apigw_.EndpointType.REGIONAL]),
+            handler=func_oauth,
             deploy=False,
+            proxy=False,
         )
+
+        item = api.root.add_resource("oauth2")
+        item.add_method("ANY", apigw_.LambdaIntegration(func_oauth))
 
         # Create APIGW Loggroup for setting retention
         LogGroup(
@@ -71,19 +71,25 @@ class SlackAppConstructsStack(Stack):
         new_deployment = apigw_.Deployment(self, f"{id}-API-Deployment", api=api)
         apigw_.Stage(
             self, f"{id}-API-Stage",
-            data_trace_enabled=False,
+            data_trace_enabled=True,
             description=f"{stage} environment",
             deployment=new_deployment,
-            logging_level=apigw_.MethodLoggingLevel.ERROR,
+            logging_level=apigw_.MethodLoggingLevel.INFO,
             metrics_enabled=True,
             stage_name=stage,
             tracing_enabled=False,
         )
 
-    def create_lambda(self, function_name: str, custom_role: iam_.Role) -> lambda_.Function:
-        if custom_role is None:
-            custom_role: iam_.Role = self.create_default_role(function_name)
+    def create_dynamodb_table(self, table_name: str) -> ddb_.Table:
+        return ddb_.Table(
+            self, table_name,
+            billing_mode=ddb_.BillingMode.PAY_PER_REQUEST,
+            partition_key=ddb_.Attribute(name="access_token", type=ddb_.AttributeType.STRING),
+            removal_policy=RemovalPolicy.RETAIN,
+            table_name=table_name,
+        )
 
+    def create_lambda(self, function_name: str, custom_role: iam_.Role) -> lambda_.Function:
         return lambda_.Function(
             self, f"{self.id}-{function_name}",
             code=lambda_.Code.from_asset(
@@ -103,7 +109,7 @@ class SlackAppConstructsStack(Stack):
             tracing=lambda_.Tracing.DISABLED,
         )
 
-    def create_immediate_response_execution_role(self, function_name: str) -> iam_.Role:
+    def create_func_oauth_execution_role(self, function_name: str, table_arn: str) -> iam_.Role:
         role_name = f"{function_name}-ExecutionRole"
         return iam_.Role(
             self, role_name,
@@ -113,13 +119,11 @@ class SlackAppConstructsStack(Stack):
                     statements=[
                         iam_.PolicyStatement(
                             actions=[
-                                "lambda:InvokeFunction",
-                                "lambda:InvokeAsync",
+                                "dynamodb:PutItem",
                             ],
                             effect=iam_.Effect.ALLOW,
                             resources=[
-                                self.func_async_worker.function_arn,
-                                self.func_sync_worker.function_arn,
+                                table_arn
                             ],
                         ),
                         iam_.PolicyStatement(
@@ -128,24 +132,13 @@ class SlackAppConstructsStack(Stack):
                             ],
                             effect=iam_.Effect.ALLOW,
                             resources=[
-                                f"arn:aws:ssm:{self.region}:{self.account}:parameter{PARAMETER_NAME}",
+                                f"arn:aws:ssm:{self.region}:{self.account}:parameter{CLIENT_ID_PARAMETER_NAME}",
+                                f"arn:aws:ssm:{self.region}:{self.account}:parameter{CLIENT_SECRET_PARAMETER_NAME}",
                             ],
                         ),
                     ]
                 )
             },
-            managed_policies=[
-                iam_.ManagedPolicy.from_aws_managed_policy_name("service-role/AWSLambdaBasicExecutionRole"),
-                #iam_.ManagedPolicy.from_aws_managed_policy_name("AWSXrayWriteOnlyAccess"),
-            ],
-            role_name=role_name,
-        )
-
-    def create_default_role(self, function_name: str) -> iam_.Role:
-        role_name = f"{function_name}-ExecutionRole"
-        return iam_.Role(
-            self, role_name,
-            assumed_by=iam_.ServicePrincipal("lambda.amazonaws.com"),
             managed_policies=[
                 iam_.ManagedPolicy.from_aws_managed_policy_name("service-role/AWSLambdaBasicExecutionRole"),
                 #iam_.ManagedPolicy.from_aws_managed_policy_name("AWSXrayWriteOnlyAccess"),